Keep layers you don't operate on
The old functionality replaces all slicer layers with the generated
ones, even if you decide to omit a layer or few. I ran into plate
adhesion issues with the pattern and wanted my initial burn-in layer to
remain as it was. This change re-concatenates the unchanged layers to
the start and end.